NCBC-1119: Correct MutationToken to include BucketRef
Motivation
----------
MutationToken and MutationState are currently unusable for N1QL queries
because MutationToken.BucketRef is always null.

Modifications
-------------
Add BucketRef to the constructor of MutationToken, and mark the old
constructor as obsolete.

Update IVBucket and VBucket to have the BucketName, which can then be used
by the various operations to build the MutationToken.

Update VBucketKeyMapper to pass the BucketName through from the
CouchbaseConfigContext to the VBucket.

Updated tests to account for the new properties and constructors.  Added
an integration test that validates that MutationTokens are created with a
BucketRef value.

Results
-------
MutationToken and MutationState will now have a valid BucketRef value for
use with QueryRequest.ConsistentWith.

Change-Id: I39ad6e38993f9fa85eec5e2045ede3f06ab63095
Reviewed-on: http://review.couchbase.org/64314
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>