NCBC-1120: Ensure UseSsl and UseEnhancedDurability are overridden when set
Motivation
----------
Fixes a bug where if UseEnhancedDurability was set to true in a
App/Web.cong and the connectionPool element was not defined, it would be
overwritten when the config initialies and would always be false.

Modifications
-------------
Updated ClientConfiguration so that the UseEnhancedDurability flag would
be set for a bucket in which no connectionPool element was defined.

Results
-------
If a bucketConfiguration is defined without a connectionPool element and
UseEnhancedDurability is true, the client will use enhanced durability.

Change-Id: Ibd4cf90473dae2c66ccab0fe339168f8bb42e27c
Reviewed-on: http://review.couchbase.org/64338
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>