NCBC-1121: RYOW ConsistentWith Doesn't Function For N1QL Queries
Motivation
----------
Fixes a bug where the scan vectors were treated as a string and not as a
JSON object resulting in the server returning: Error #: 1070
Message: scan_vectors has to be of type map of strings to vectors

Modifications
-------------
Removed call to serialize the result MutationState object, which converts
it to a string instead of a JSON object.

Results
-------
The scan vectors will be properly serialized as JSON and included as part
of the request body.

Change-Id: Icbb43775c4267b2b44d7acfefd711086fd5900f6
Reviewed-on: http://review.couchbase.org/64437
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>