NCBC-1122: Add config flag to ignore TLS FQDN validation
Motivation
----------
If you bootstrap with an IP (and not an FQDN) and are using TLS, the
client will fail with a RemoteCertificateNameMismatch because the
certificates name doesn't match the IP. As a work-around for test
environments, we can add a flag that ignores this error and allows the
client to not fail when establishing the TLS connection.

Modifications
-------------
Added static IgnoreRemoteCertificateNameMismatch config property to
ClientCOnfiguration and instance properties to CouchbaseClientDefinition and
CouchbaseClientSection. Added additional logic to ignore
RemoteCertificateNameMismatch during authentication of connections.

Results
-------
If you are using IP's and TLS/SSL, you can enable
IgnoreRemoteCertificateNameMismatch so that these validation errors are
ignored.

Change-Id: I5c0afc25c9a41814fbabc555101513a67dc254fb
Reviewed-on: http://review.couchbase.org/64503
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>