NCBC-1117: HttpServerConfig Doesn't Validate Server Certificate
Motivation
----------
When calling HttpServerConfig.Initialize, the server certificate is not
validated; all certificates are accepted.

Modifications
-------------
Added certificate validation and an optional override if
RemoteCertificateNameMismatch is generated (typically by bootstrapping
with an IP instead of a FQDN).

Result
------
When bootstrapping, if SSL is enabled the SDK will validate the
certificate.

Change-Id: Ibe0b835629de08ce7e3d8734e036c24b05dda2ba
Reviewed-on: http://review.couchbase.org/64504
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>