NCBC-1126: BinaryToJsonTranscoder implementation
Motivation
----------
In some cases users of the older SDK (pre-2.X) may be using a custom
transcoder which encodes using Binary serialization. Since all 2.X clients
using JSON serialization, there is a backwards-compatibility issue. To
help with these cases, this transcoder reads in binary format and writes
in JSON. Note that most users will not need this transcoder!

Modifications
-------------
Added a transcoder which uses Binary serialization for decoding and JSON
serialization for encoding.

Change-Id: I5c9269aaf26ce63a87a63ad6ffac210bea9006f7
Reviewed-on: http://review.couchbase.org/64627
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>