NCBC-1111: Remove exp. backoff for NMV and replace with static 100ms backoff
Motivation
----------
Removes the exponential backoff in place when a NMV and replaces it with a
default 100ms backoff.

Modifications
-------------
Removed th exponential backoff for NMV in the RequestExecutorBase and
CouchbaseRequestExecutor classes for K/V operations. Added a
VBucketRetrySleepTime property to ClientConfiguration,
CouchbaseClientSection and CouchbaseClientDefinition configuration
classes. Added basic unit tests.

Results
-------
If a NMV is encountered, the client will wait 100ms to give time for the
server to work it state out, it will then retry the operation at 100ms
until it times out or succeeds.

Change-Id: Ie74c1aa8bd61473063a6a4b65c85a885a5137b51
Reviewed-on: http://review.couchbase.org/64701
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>