NCBC-1131: Regression causes HttpStreamingProvider to auth invalid password
Motivation
----------
Fixes a regression where an invalid password for a bucket would allow a
bucket to be created by OpenBucket(), however subsequent calls on one of
the API's (document, n1ql, etc) would fail authentication.

Modifications
-------------
Added AuthenticationHeaderValue directly to the HTTP headers in
AuthenticatingHttpClientHandler, rather than using the Credentials
property.  Updated the int. test so that it always fails if a
AuthenticationException is found in the InnerExceptions list in
AggregateException thrown by client when bootstrapping fails.

Result
------
The client will fail-fast during bootstrapping if the bucketname or
password are incorrect.

Change-Id: I2a2ecbd14fc17d5d13999847a0dbfde585497b59
Reviewed-on: http://review.couchbase.org/64749
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>