NCBC-1133: Remove unused IOperationResult.cs file
Motivation
----------
When converting to a .xproj/project.json project system, all files are
included.  The extraneous IOperationResult.cs is therefore being included
in the build.

Modifications
-------------
Remove the defunct file.

Results
-------
Fixes the issue in the new project file format once we complete the
conversion.  This file was not included in the current .csproj, so the
current build is unaffected.

Change-Id: I1265e62bd7618f51268767bae90ecbc24b518d7d
Reviewed-on: http://review.couchbase.org/64750
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>