NCBC-1135: Use Endpoint instead of Uri for SslStream.Authenticate
Motivation
----------
If UseSsl is enabled, the client will try to use the Uri.Host value to
Authenticate which results in a
SslPolicyErrors.RemoteCertificateNameMismatch error. Instead, pass the IP
value of EndPoint which correctly validates.

Modifications
-------------
Changed SslConnection.Authenticate() method to use the IP value of the
EndPoint property.

Result
------
Client correctly authenticates FQDNs when UseSsl is enabled.

Change-Id: Ia286707685dbe78a00a763b5425a5e4ac6d33cda
Reviewed-on: http://review.couchbase.org/64759
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>