NCBC-1137: Auth error causes BootstrapException to be thrown
Motivation
----------
This appears to be a regression where the AuthenticationException is not
rethrown when caught, thus BootstrapException is thrown is since
HttpServerConfig.DownLoadConfig returns false.

Modifications
-------------
Added throw clause in catch block of DownLoadConfigs - this causes the
real exception to be bubbled up.

Result
------
If an auth error occurs when bootstrapping, an AuthenticationException
will be thrown instead of a more generic BootstrapException.

Change-Id: I55024bf8037eebfc37d92196627761be32f4a5e9
Reviewed-on: http://review.couchbase.org/64828
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>