NCBC-1147: MutationToken uses wrong bucket - impacts RYOW
Motivation
----------
When UseEnhancedDurability is enabled, default bucket is assigned to the
BucketRef field of the MutationToken which causes RYOW to not work
correctly.

Modifications
-------------
Remove hard-coded default bucket name and use passed in bucket name from
ctor. Add unit and integration tests.

Results
-------
Correct BucketRef is used for MutatationToken's so that RYOW works as
expected.

Change-Id: I0d804e5764ed09672ffe0806cccd95323f614883
Reviewed-on: http://review.couchbase.org/65165
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>