NCBC-1143: Subdocument Exists always returns true
Motivation
----------
Fixes a bug where IDocumentFragment.Exists will return true if the path
does not exist for the document.

Modifications
-------------
Change DocumentFragment.Exists to first check for the existence of the
path spec and then to check to see if the status is Success or not.

Result
------
If the path does not exist in the document, false will be returned.

Change-Id: I7c13a80e1fd31961c6ac20f061ec7f77d6ed60d0
Reviewed-on: http://review.couchbase.org/65275
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>