NCBC-1158: NotImplementedException when rebalance with Sub-Doc API
Motivation
----------
This fixes a bug where Sub-Document API operations would fail with an
NotImplementedException during a rebalance because the base Clone() method
would be called which is not implemented.

Modifications
-------------
Added Clone() implementations to each Sub-Document API operation.

Results
-------
NotImplementedExceptions are not thrown by Sub-Document API methods during
a rebalance.

Change-Id: Ic82edb7eedf7eed69019c9ec4f682862b5767b1d
Reviewed-on: http://review.couchbase.org/65624
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>