NCBC-1160: Add option to enable N1QL timing logging
Motivation
----------
Currently there is no way to measure the time taken for N1QL queries to
complete. Ideally we would want to record the time taken end-to-end from
the client, the cluster elapsed time and the raw query that was submitted.

Modifications
-------------
Introduce a QueryTimer construct similar to OperationTimer that can be used
to wrap the execution of queries in QueryClient.ExecuteQueryAsync<T>. After
execution the timer instance receives the QueryResult.Metrics.ElaspedTime
so it can properly format the log entry.

Includes supporting configuration in ClientConfiguration and
CouchbaseClientSection. This adds the setting 'EnableQueryTiming' in
web/app.config under the Couchbase config section. To enable logging both
'EnableQueryTiming' must be set to true and the Log appender must be set to
Info or finer.

Test cases haven been added to ensure:
- configuration permutations
- no ElapsedTime is set
- Log components
- QueryRequest is null
- QueryRequest statement is missing or null

Results
-------
With the appropriate configuration log entries similar to the following are
logged:
2016/07/10 01:06:04:107 [INFO]  mylogger - Query Timing: 4.197984ms |
3.764454ms | "SELECT * FROM default"

Further Enhancements
--------------------
TimingFactory
I have used the OperationTimer as a guide to try and get a feel for how timing
has been dealt with in the Core namespace. My QueryTimer implementation has
followed fairly closely however I did not implement a TimingFactory for
QueryTimer as I believed this to be going a little too far for this small
project and was adding more complexity.

Setting ElapsedTime during execution of request
I don't really like the way I have had to set the ServerElapsedTime. On review
it might have been better to build a structure that given a LogStore,
QueryRequest and QueryResult it would write the appropriate log entry but
would still need a StopWatch to measure the time taken for the HTTP request.

If I had more time
I would have liked to make a cleaner wrapper structure for the Query Request /
Result objects that could offer easier access to the required properties to
form the log entry. I also feel I may have gone down the wrong path with
trying to replicate the OperationTimer implementation and made it work for
query timings, another approach would have been better.

Notes
-----
There are a number of broken tests however I do not believe they are related
to my proposed changes.

Change-Id: Ie0c43539d692014bcb8dde86f5626a8cc9b75edd
Reviewed-on: http://review.couchbase.org/65641
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>