NCBC-1162: Make Observe fail-fast when document mutation is encountered
Motivation
----------
While performing an observe operation on a document, if another thread
updates the document (changing the CAS), then the observe operation should
fail with a DocumentMutationException.

Modifications
-------------
Added DocumentMutationException which will cause control to break from
the observe polling loop if a mutation is encountered. Added
DocumentMutationDetected enum value to ResponseStatus. Override ToString
in Observestate. Update CouchbaseRequestExecutor to handle
DocumentMutationException and return back the proper error to the
application. Added additional logging and properly return false whenever a
failure state is encountered in CouchbaseRequestExecutor.

Results
-------
If a document mutation is encountered while doing an observe poll, the SDK
will fail-fast and return back the correct response code. This will reduce
latency, since the SDK will not wait for the operation to timeout because
its already in a failure state.

Change-Id: I85908a12821f7651e0e8ccfa8c3b28735cf5a384
Reviewed-on: http://review.couchbase.org/66018
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>