NCBC-1157: Queries fail if DefaultSerializer.ContractResolver is null
Motivation
----------
One of the DefaultSerializer constructors receives two instances of
JsonSerializerSettings to mange serialization and deserialization tasks.
However, if the ContractResolver property is null when trying to execute
GetMemeberName a NullException is thrown (line 209).

It would be better to use the JsonConvert.DefaultSettings ContractResolver
if available otherwise fallback to the DefaultContractResolver.

Modifications
-------------
Check and default both serializationSettings and deserializationSettings
ContractResolver properties to use the JsonConvert.DefaultSettings
ContractResolver, falling back to Default DefaultContractResolver if
unavailable.

Adds supporting tests to ensure:
- DefaultContractResolver is used if settings.ContractResolver is null and
  JsonConvert.DefaultSettings is also null
- JsonConvert.DefaultSettings ContractResolver is used if
  settings.ContractResolver is null and
  JsonSettings.DefaultSettings.ContractResolver is not null

Results
-------
Both serializationSettings and deserializationSettings ContractResolver
properties are set use either the JsonConvert.DefaultSettings
ContractResolver or DefaultContractResolver if they are
null during construction of DefaultSerializer.

Executing GetMemberName after creating DefaultSerializer with null
ContractResolver does not throw NullException.

Change-Id: Ibe812da8c99e343eb4d50c957eeddb3edead66ba
Reviewed-on: http://review.couchbase.org/66108
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>