NCBC-1157: Buckets get opened with a password provided for another in the config .NET SDK
Motivation
----------
When opening buckets using ConfigProviderBase and there is no previously
opened bucket a new BucketConfiguration is created, if there is an existing
configuration available it is cloned to promote reusing Cluster settings; such
as Server URLs. However one of the fields that is copied is the bucket
password which is not intended as it's possible to have different passwords
per bucket.

Modificiations
--------------
Default the password field for subsequent buckets to an empty string.

Added test to prove the initial bucket's password is not used on
subsequent bucket configurations and instead the password is an empty
string.

Results
-------
When subsequent buckets are opened the original buckets password is no longer
copied onto the new bucket configuration and is instead an empty string.

Change-Id: I84bbe1d0c206078c0b0a6040de4d9625c2c0e51a
Reviewed-on: http://review.couchbase.org/66177
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>