NCBC-1164: Document mutation incorrectly indicated using CAS-based observe
Motitivation
------------
Fixes a bug where the SDK checks the CAS, which is always zero when Not Found
or Not Persisted, and incorrectly flags the operation as a document
mutation.

Modifications
-------------
Changed logic to check the CAS and the persisted state before throwing
DocumentMutationException. Fixed a typo in the TestConfiguration class.
Added unit tests.

Results
-------
SDK will correctly perform observe only failing if the CAS differs from
the observe state.

Change-Id: I68fd5ed6ffb5868f746f14522072cb1bc4b67bda
Reviewed-on: http://review.couchbase.org/66250
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>