NCBC-1163: MUX IO - OperationTimeout when Enhanced Durability is enabled
Motivation
----------
Fixes a bug where the where threads may deadlock waiting for an operation
to complete when using the MultiplexingConnection and
MultiplexingIOService (MUX).

Modifications
-------------
Added Task.Run when state.Complete() is called so that it executes on the
threadpool and not on the current thread.

Result
------
OperationTimeout is no longer encountered when running several operations
asynchronously.

Note: a further improvement would be to rewrite the MUX implementation so
that it's purely asynchronous and not hybrid.

Change-Id: Id6b94985f7c93be3fc87f5355f0d74b7e0434351
Reviewed-on: http://review.couchbase.org/66275
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>