NCBC-1169: Make exception behavior more like Java SDK
Motivation
----------
This commit makes the .NET sdk return/throw similarly named exceptions
for various failure states:
DocumentAlreadyExistsException - an attempt to insert a doc that exists
DocumentDoesNotExistException - an attempt to replace a doc that does not
exist
TemporaryLockFailureException - an attempt to retrieve a locked doc
CasMismatchException - a doc was modified resulting in a CAS failure
Previously, the IOperationResult.Exception field was null on these cases,
however the Status field indicated the failure.

Modifications
-------------
Added new exception classes. Updated the RequestExecutor impl. to use the
SetException() method for each specific failure state.

Result
------
The .NET and Java SDK will return/throw similar exceptions for the same
common set of failure states.

Change-Id: I523bf6dc60d3cd3811ca8c34ee0a7724b942a2d6
Reviewed-on: http://review.couchbase.org/66398
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>