NCBC-1170: Make Insert and Upsert return the Document and Id
Motivation
----------
This is another commit to make the .NET SDK more consistent with the Java
SDK's behavior. Now, when you insert, upsert, replace on a document (using the Document<T>
API), the DocumentResult will contain the new or updated document and it's
id. Since the CAS is included as well, this makes it easier to reuse
Document<T> objects.

Modifications
-------------
Updated the DocumentResult<T> classes ctor to take a OperationResult and a
Document - the document being the recently inserted or modified document.
Ensured that the document Id was assigned back to the DocumentResult<T>
object after a mutation. Removed the Id parameter from the CTOR, since it
now comes from the OperationResult.

Results
-------
After an insert (or other mutation), the DocumentResult<T> insance will
contain a copy of the latest document's content. All OperationResult and
derived classes will contain the Id of the document.

Change-Id: I0990fadcf8a0f794477e53faa95a50aa6932321a
Reviewed-on: http://review.couchbase.org/66478
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>