NCBC-1168: Resolve compiler warnings
MOTIVATION
----------
There are ~100 compiler warnings when building the solution. There are
made up from a mix of using depreciated enums & classes, unused variables
and others. We should not have any compiler warnings.

MODIFICATIONS
-------------
- Use #pragma keywords to ignore internal depreciated class and enum
  usage
- Remove usunued variables
- Remove > -1 checks against uint properties
- Convert backing fields to auto properties
- Stored the output of TryRemove out param into a local variable instead
  of the variable that was under lock
- Replace usage of ManualByteConverter with DefaultConverter
- Removed overriden Id property in DocumentFramgent`.cs
- Make ObserveSeqno.cs Id property override instead of virtual

RESULT
------
All related warnings are no longer present during build.

NOTE: The was one unusual build warning in Couchbase.IO.ConnectionPool.cs
line 275 that needed a code change. The warning is because the object used
in the synchronisation lock command is being modified during the lock. I
fixed this by assigning the output param from the
ConcurrentDictionary.TryGet to a local variable instead of to the main conn
variable.

Change-Id: Iceb6d21acff0b113b8c99ddd3a091a773423e556
Reviewed-on: http://review.couchbase.org/66723
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>