NCBC-1171: .Net client should send HELO command during bootstrap
MOTIVATION
---------
During bootstrap the SDK should send a HELO command to the server.
Currently this is only performed if UseEnhancedDurability is enabled in
the config. The Hello command is supported from Server version 4.0+.

MODIFICATIONS
-------------
Execute a HELO command during bootstrap in both PooledIOService.cs
and MultiplexingIOService.cs even when UseEnhancedDurability is set to
false.

Add new constructor to Hello.cs that defaults the key and add a private
const to store the key format.

Update Hello usage to use the new constructor.

Removed unused EnableEnhancedDurabilityAsync method. We cannot use an
async method because it needs to be executed during a synchronisation
lock which you can't use together.

Added logging for if the Hello command fails.

RESULT
------
A HELO command is sent to the server whether UseEnhancedDurability is
set or not when bootstrapping the SDK connection.

NOTE: The hello commands continued to succeed even on server version
3.1.6 running under docker from tag couchbase:3.1.6.

Change-Id: If7fd2209271287165990711ecb0ea07bb3ea1fb8
Reviewed-on: http://review.couchbase.org/66691
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>