NCBC-1141: When Id is not present in doc fail-fast
MOTIVATION
----------
Whenever mn insert, replace, append or prepend command is executed without a
key property an exception is thrown during serialization. It would be
better to fail-fast if the key was null or empty.

Whenever an operation that requires a key is executed without a valid
key an exception is thrown during serialization. It would be better to
fail-fast and throw an exception when the operation is created with an
invalid key.

MODIFICATIONS
-------------
Add new KeyException that sets the message to "Key cannot be null or
empty.".

Thorw KeyException during construction if key is null or empty for:
- Get, GetL, GetQ, GetQ and GetT
- Hello
- Add, Append, Decrement, Delete, Prepend, Replace and Set
- Observe
- ReplicaRead
- Touch
- Unlock

- All Subdocument operations

Added tests to prove a null or empty key throws for each operation type
correctly.

Also added local.config to .gitignore so we can have local appSettings
that don't need to be added to source control.

RESULT
------
If an insert command is executed without a valid key the operation now
fails quickley before trying to serialize the operation.

Change-Id: I8a155c6909204ac60059b1c4ce3becddd3153510
Reviewed-on: http://review.couchbase.org/66725
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>