NCBC-1132: OperationResult will return ClientFailure when an Auth failure
MOTIVATION
----------
When a SASL authentication error occurs the error message does not explain
the problem and the ResponseStatus is incorrectly set as ClientFailure.
A more appropriate message and ResponseStatus would make debugging
easier.

MODIFICATIONS
-------------
Update the message when an auth error occurs to be more descriptive, eg
"Authentication failed for bucket '<bucket>'".

Catch AuthenticationExceptions independently from Socket or general
Exceptions to set the ResponseStatus to AuthenticationError during
synchronous Execute methods.

Set the ResponseStatus appropriately when handling async exceptions.

Added tests to ensure message and ResponseStatus is set as expected.

Also tidied up a couple field declarations in MultiplexingIOService.cs
and PooledIOService.cs.

Add message to ExceptionUtils.cs and use when constructing
AuthenticationException in both PooledIOSerivce and
MultiplexingIOService.

RESULT
------
The ResponseStatus is correctly set to AuthenticationError and a more
descriptive error message is returned when SASL authentication fails.

Change-Id: I2b3f450cc78d68c86116e75d4c2d8e20c7f04395
Reviewed-on: http://review.couchbase.org/66730
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>