NCBC-1062: Add cluster-level authenticator interface
Motivation
----------
In the current implementation of client libraries, authentication is
provided on an as-needed basis.  This means that a password is provided to
the OpenBucket method, and this password is used throughout the clients to
authenticate to various cluster services.  In addition, when a new manager
object is created for a Cluster or Bucket, the credentials are provided at
creation time.  The addition of cluster-level queries, which currently
rely on the credentials of multiple buckets to execute, as well as the
inclusion of more advanced form of cluster level authentication such as
RBAC or CBAUTH have made it clear that an improved client authentication
model is necessary.

Modifications
-------------
Added IClusterCredentials interface and new method for passing in the
IClusterCredential impl. to the ICluster object. Internally, updated all
API's to use the IClusterCredentials if they are present, otherwise
default back to the standard authentication scheme where the password is
passed into IBucket.OpenBucket.

Results
-------
The SDK will use the passed in IClusterCrendentials impl for
authentication if present, falling back to the legacy bucket level
authentication if not present.

Change-Id: I7e062667bc3781c4f65872408fd0dca2e9b77091
Reviewed-on: http://review.couchbase.org/66749
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>