NCBC-1182: When ConfigThreadState shuts down log INFO instead of ERROR
Motivation
----------
When the ConfigThreadState class is disposed, it currently (< 2.3.6)
throws and logs and ObjectDisposedException as an ERROR This is misleading
because the exception is thrown on purpose to break from the infinite loop
and stop listening for config changes from the server. This commit catches
the exception and logs it as an INFO with the following message: "The
config listener has shut down."

Modifications
-------------
Added catch for AggregateException and check it for
ObjectDisposedException. Log message to INFO, remove redundent catch
blocks.

Result
------
When the config listener thread is terminated, it will be logged similar
to: 2016-08-16 16:33:50,309 [17] INFO
Couchbase.Configuration.Server.Providers.Streaming.ConfigThreadState - The
config listener has shut down.

Change-Id: I54a18fd9fb9534d6ff70322a2a797e303a23b464
Reviewed-on: http://review.couchbase.org/66821
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>