NCBC-1184: Fixed Document<T> missing Token
Motivation
----------
For some operations updated by changes for NCBC-1170, the Document object
is missing the Token value, it is only present on the DocumentResult
object.

Modifications
-------------
Ensure that the Token is copied from the OperationResult to the Document
as the DocumentResult is being constructed, consistent with previous
behavior.

Update affected integration tests slightly to get better test results,
since enhanced durability must be enabled on the test buckets for the
tests to succeed.

Results
-------
Document objects now have Tokens again, and affected tests are passing.

Change-Id: I09f6054e9782fbdf430f6bbf927d81e91c20ae03
Reviewed-on: http://review.couchbase.org/66928
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>