NCBC-1192: Remove NewtonSoft dependency from ViewResult
Motivation
----------
The ViewResult class has dependencies on JsonPropertyAttribute which is
from the NewtonSoft API; this patch removes the dependency and replaces it
with a standard .NET class.

Modifications
-------------
Remove JsonProperty attributes, add POCO with fields that match naming
conventions of the View API response, and map the response to the exposed
ViewDataRow POCO.

Results
-------
There is no longer a hard dependency on NewtonSoft in the ViewResult
class.

Change-Id: I9086d569505383244ac6756e0942a02419faf1fa
Reviewed-on: http://review.couchbase.org/67198
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>