NCBC-1060: Add Data Structure support
Motivation
----------
It is common for developers to use common data structures as a model for
their document and to perform the operations defined by a data structure
on that document. This commit adds four (4) different data structure
abstractions whose items are persisted to and stored in Couchbase:
CouchbaseSet, CouchbaseList, CouchbaseDictionary and CouchbaseQueue. The
internal operations are optimized where they can be by use Sub-Document
commands, in other places where the correct behavior could not be achieved
using the Sub-Doc API, regular Memcached operations (GET, SET, etc) are
used. Since the Sub-Doc API is used in most places, the current
implementation will only work with Couchbase 4.5 or greater. Also note,
since there is network and other latency involved, the Big O
characteristics for these data structures may not be realized.

IMPORTANT NOTE: this API should be considered "volatile". While the
public methods and names will likely not change, the internal
representation will likely change as new features are added to the Sub-Doc
API and as API matures.

Modifications
-------------
There is a new namespace called Couchbase.Collections. In that namespace
there are new classes CouchbaseSet, CouchbaseDictionary, CouchbaseList and
CouchbaseQueue classes and associated tests.

Result
------
AS a developer, you can now treat Couchbase documents as common data
structures, performing familiar operations such as Enqueue, Dequeue, etc
on the document.

Change-Id: Iad7a9f39041d39fdfe38f6d249b5c7da9844b746
Reviewed-on: http://review.couchbase.org/67687
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>