NCBC: Unable to submit Compound FTS query
MOTIVATION
----------
When creating either a conjuncts or disjuncts FTS query and submitting
to the server an exception of '"Can not add Newtonsoft.Json.Linq.JArray
to Newtonsoft.Json.Linq.JObject."'. This generated during the
serialisation process before sending to the server.

The SDK should generate valid JSON to be sent to the server to execute
the query.

MODIFICATIONS
-------------
Update 'new JArray' implementations to be wrapped inside a JProperty
when executing Export for ConjuctionQuery and DisjunctionQuery.

Added unit tests to prove valid JSON is returned from the Export method
call for both query types.

RESULT
------
Valid JSON is returned when executing the Export method for both query
types.

Change-Id: I1446a250df838cd86290fac477059c16dc5e5012
Reviewed-on: http://review.couchbase.org/68299
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>