NCBC-1214: BooleanQuery should only require one sub-query, not one of each type
MOTIVATION
----------
The FTS BooleanQuery is made up of three sets of sub-queries; must,
must_not and should. Only one of these collections must contain an item
to be valid to send to the server.

Currently, all three collections must have an entry to be valid or an
exception is thrown.

Also - the BooleanQuery also suffers from the same JArray issue as found
in NCBC-1213.

MODIFICATIONS
-------------
When generating the query JSON as part of the Export method, only
require one of the sub-query collections to contain an item.

Update JArray implementations to be wrapped in a JProperty so the JSON
serialiser can generate the JSON successfully.

Added unit tests for both of the above cases.

RESULT
------
Only one sub-query in one of the collections is required for the
BooleanQuery to be deemed valid and the Export method successfully
generates JSON.

Change-Id: Ic491b6b4dd24cda734cc0c4b36d05930b23460df
Reviewed-on: http://review.couchbase.org/68316
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>