NCBC-1224: part 2 - IBucket level data structure methods
Motivation
----------
This is part two (2) of the Data Structures feature; the first one is
NCBC-1060 and is an API of Collections that are persisted to Couchbase.
This commit includes a new bucket level API for various data structure
methods for lists, sets and hashmaps.

Modification
------------
Add signatures for common data operations to be peformed on a data
structure: set, add, remove, etc.

Results
-------
In addition to the traditional IBucket key-value methods, there are now
methods which operate within the document on specific data structures:
arrays, sets and hashmaps. You use these like any other K/V methods,
except that they operate generally within the document and not on the
entire document (with some exceptions). Note that these methods are only
supported by Couchbase (peristent) buckets.

Change-Id: I31a4ffa953fa2be0f4196163b39d042d0862a85a
Reviewed-on: http://review.couchbase.org/68691
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>