NCBC-1221: Make OpenBucket throws Auth exception when cred not provided
Motivation
----------
When a bucket is attempted to be opened using an Authenticator, if the
BucketCredentials do not exist for that bucket, throw a
AuthenticationException. Previously a KeyNotFoundException was thrown by
the dictionary, which is not very useful.

Modifications
-------------
Change code to throw an AuthenticationException if the bucket does not
exist in the BucketCrendentials map. Update test.

Results
-------
The SDK will throw a AuthenticationException when a bucket is attempted to
be opened and it's credentials are not found.

Change-Id: Iff2b0f900e5724e5dfdd9fc8d1fba32ae9ebdcf9
Reviewed-on: http://review.couchbase.org/68786
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>