NCBC-1194: Bucket.Exists() returns False for valid key with non-standard chars
MOTIVATION
----------
Executing an Exists or ExistsAsync for a document that includes Swedish
characters always returns false, even if the document exists. This is
because the Write method on the ObserverOperation is not calculating the
byte array correctly and the buffer was too short which caused an
exception and a default value of False was returned.

MODIFICATIONS
-------------
Use the OperationBase method to calculate the key length in bytes, which
converts the Key to UTF8 first. This value is then used to calculate the
buffer array length.

Added test to prove non-standard UTF8 characters can be used in document
key.

RESULT
------
ObesrverOperations can now correctly calculate the byte buffer array for
keys with non-standard UTF8 characters which allows Exists and
ExistsAsync to execute.

Change-Id: Ie156b0b94b52b4d4c01e0351b9a37ce3621c6dea
Reviewed-on: http://review.couchbase.org/68822
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: buildbot <build@couchbase.com>